requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyairtable>=2.3.0
rapidfuzz>=3.6.0
python-dotenv>=1.0.0
